Creates Board Controller
